requirements.txt
Django==1.8.1
Markdown==2.6.1
Pillow==2.8.1
pydocs==0.1
pytz==2015.2
redis==2.10.3
celery==3.1.17
django-celery==3.1.16
django-filter==0.9.2
djangorestframework==3.1.1
django-extensions==1.5.2
django-push-notifications==1.2.1
django-debug-toolbar==1.3.0
docutils==0.12
gunicorn==19.3.0
uWSGI==2.0.10
